feat: 함수 분리
