+++ src/hooks/useCalcDiscount.ts
@@ -2,15 +2,9 @@ import {ChangeEvent, useEffect, useState} from 'react';
 import {CouponCode, CouponType} from '../type/coupon';
 import {useShowError} from '../provider/errorProvider';
 import {getCoupons} from '../api/coupon/getCoupon';
-import {useSelectedItems} from '../provider/cartItemsProvider';
-import {
-  discountByBOGO,
-  discountByFIXED5000,
-  discountByMIRACLESALE,
-} from '../feature/calcCouponPrice';
 
-export const useCalcDiscount = () => {
-  const selectedItems = useSelectedItems();
+export const useCoupons = () => {
+  const [coupons, setCoupons] = useState<CouponType[]>();
   const [isCouponChecked, setIsCouponChecked] = useState<
     Record<CouponCode, boolean>
   >({
@@ -20,7 +14,6 @@ export const useCalcDiscount = () => {
     MIRACLESALE: false,
   });
   const showError = useShowError();
-  const [coupons, setCoupons] = useState<CouponType[]>();
 
   useEffect(() => {
     const getCouponData = async () => {
@@ -55,34 +48,5 @@ export const useCalcDiscount = () => {
     }));
   };
 
-  const calcDiscount = (orderPrice: number) => {
-    let discountResult = orderPrice;
-    const checkedCoupons = Object.keys(isCouponChecked).filter(
-      (key) => isCouponChecked[key as keyof typeof isCouponChecked]
-    );
-
-    // 30%할인, 2+1 할인 쿠폰 적용되어 있을 때
-    if (
-      checkedCoupons.includes('MIRACLESALE') &&
-      checkedCoupons.includes('BOGO')
-    ) {
-      return Math.min(
-        discountByBOGO(discountByMIRACLESALE(discountResult), selectedItems),
-        discountByMIRACLESALE(discountByBOGO(discountResult, selectedItems))
-      );
-    }
-
-    if (checkedCoupons.includes('MIRACLESALE'))
-      discountResult = discountByMIRACLESALE(discountResult);
-
-    if (checkedCoupons.includes('BOGO'))
-      discountResult = discountByBOGO(discountResult, selectedItems);
-
-    if (checkedCoupons.includes('FIXED5000'))
-      discountResult = discountByFIXED5000(discountResult);
-
-    return discountResult;
-  };
-
-  return {isCouponChecked, coupons, handleCouponsChecked, calcDiscount};
+  return {isCouponChecked, coupons, handleCouponsChecked};
 };
